+++ src/core/prelude/object/convert/CHANGELOG.md
@@ -11,13 +11,16 @@ Changelog
 
 ## v3.54.0 (2021-07-09)
 
-#### :bug: Bug Fix
+#### :boom: Breaking Change
 
-* Fixed `throttle` callback invoking
+* The module has been renamed from `parse` to `convert`
+
+#### :rocket: New Feature
+
+* Added `Object.trySerialize`
 
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature
 
 * Added tests
-* Added `single` parameter to `throttle`